+++ docs/modules/documentation/containers/search-input/examples/search-input-dynamic-example.component.ts
@@ -11,23 +11,22 @@ export class SearchInputDynamicExampleComponent {
     selected: string;
 
     dropdownValues = [
-        { text: 'Apple' },
-        { text: 'Banana' },
-        { text: 'Kiwi' },
-        { text: 'Strawberry' },
-        { text: 'Tomato' },
-        { text: 'Pineapple' }
+        { text: 'Apple', callback: () => {alert('Apple Clicked!')} },
+        { text: 'Banana', callback: () => {alert('Banana Clicked!')} },
+        { text: 'Kiwi', callback: () => {alert('Kiwi Clicked!')} },
+        { text: 'Strawberry', callback: () => {alert('Strawberry Clicked!')} },
+        { text: 'Tomato', callback: () => {alert('Tomato Clicked!')} },
+        { text: 'Pineapple', callback: () => {alert('Pineapple Clicked!')} }
     ];
-
-    dropdownSubject = this.dropdownValues;
-
-    modelChange(event: any): void {
-        this.searchTerm = event;
-        this.dropdownSubject = this.dropdownValues.filter(value =>
-                value.text.toLocaleLowerCase().includes(this.searchTerm.toLocaleLowerCase()));
-    }
 
     selectItem(event: any) {
         this.selected = event.text;
+    }
+
+    customFilter(content: any[], searchTerm: string): any[] {
+        const search = searchTerm.toLocaleLowerCase();
+        return content.filter(item =>
+            item.text.toLocaleLowerCase().startsWith(search)
+        );
     }
 }