+++ library/src/lib/search-input/search-input.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FdSearchPipe, SearchInputComponent } from './search-input.component';
+import { SearchInputComponent } from './search-input.component';
 import { PopoverModule } from '../popover/popover.module';
 import { MenuModule } from '../menu/menu.module';
 import { FormsModule } from '@angular/forms';
 
 @NgModule({
-    declarations: [SearchInputComponent, FdSearchPipe],
+    declarations: [SearchInputComponent],
     imports: [CommonModule, PopoverModule, MenuModule, FormsModule],
-    exports: [SearchInputComponent, FdSearchPipe]
+    exports: [SearchInputComponent]
 })
 export class SearchInputModule {}