+++ library/src/lib/time-picker/time-picker.component.ts
@@ -1,4 +1,4 @@
-import { Component, forwardRef, Input, OnInit, ViewChild } from '@angular/core';
+import { ChangeDetectorRef, Component, forwardRef, Input, OnInit, ViewChild } from '@angular/core';
 import { TimeObject } from '../time/time-object';
 import { TimeComponent } from '../time/time.component';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
@@ -47,18 +47,6 @@ export class TimePickerComponent implements ControlValueAccessor, OnInit {
     isOpen: boolean;
 
     placeholder: string;
-
-    readonly POPOVER_OPTIONS: PopperOptions = {
-        placement: 'bottom-start',
-        modifiers: {
-            preventOverflow: {
-                enabled: false
-            },
-            hide: {
-                enabled: false
-            }
-        }
-    };
 
     onChange: Function = () => {};
     onTouched: Function = () => {};
@@ -144,6 +132,7 @@ export class TimePickerComponent implements ControlValueAccessor, OnInit {
                 this.child.displayedHour = null;
                 this.child.period = 'am';
                 this.child.oldPeriod = 'am';
+                this.onChange(this.time);
             }
         } else if (this.meridian) {
             if (this.displaySeconds) {
@@ -178,6 +167,7 @@ export class TimePickerComponent implements ControlValueAccessor, OnInit {
                 this.child.displayedHour = null;
                 this.child.period = 'am';
                 this.child.oldPeriod = 'am';
+                this.onChange(this.time);
             }
         }
     }
@@ -225,6 +215,11 @@ export class TimePickerComponent implements ControlValueAccessor, OnInit {
         return retVal;
     }
 
+    timeFromTimeComponentChanged() {
+        this.cd.detectChanges();
+        this.onChange(this.time);
+    }
+
     registerOnChange(fn: (time: TimeObject) => void): void {
         this.onChange = fn;
     }
@@ -243,4 +238,6 @@ export class TimePickerComponent implements ControlValueAccessor, OnInit {
         }
         this.time = time;
     }
+
+    constructor(private cd: ChangeDetectorRef) {}
 }